requirements.txt
asgiref==3.7.2
cffi==1.15.1
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
